Granularily disable PCRE2 JIT on macOS ARM (backported to Qt 5)
2f8df4d1a87df9dd67913b016171dac20839d16e disabled the PCRE2 JIT on macOS
for ARM, but did so based on the build system architecture configure
test. When doing a universal build these configure tests are based on
the primary architecture (x86_64 in our case), which means we ended up
still enabling the JIT.

We work around it by passing the PCRE2_DISABLE_JIT define only when
building the arm64 slice.

Change-Id: If2975d270507e6841f812cac3cd8e565bbf0f7d7